refactor(agents): split model compatibility helpers
