+++ src/agents/embedded-agent-runner/model.ts
@@ -49,6 +49,13 @@ import {
 } from "../sessions/index.js";
 import { discoverCachedAgentStores } from "./model-discovery-cache.js";
 import {
+  mergeModelCompat,
+  mergeModelMediaInput,
+  resolveConfiguredFallbackReasoning,
+  resolveConfiguredModelReasoning,
+  resolveMergedConfiguredModelReasoning,
+} from "./model.compat.js";
+import {
   buildInlineProviderModels,
   type InlineProviderConfig,
   normalizeResolvedTransportApi,
@@ -390,29 +397,6 @@ function normalizeTransportBaseUrl(baseUrl: unknown): string | undefined {
 
 function resolveProviderRequestTimeoutMs(timeoutSeconds: unknown): number | undefined {
   return finiteSecondsToTimerSafeMilliseconds(timeoutSeconds, { floorSeconds: true });
-}
-
-function mergeModelMediaInput(
-  base: ModelMediaInputConfig | undefined,
-  override: ModelMediaInputConfig | undefined,
-): ModelMediaInputConfig | undefined {
-  if (!base) {
-    return override;
-  }
-  if (!override) {
-    return base;
-  }
-  return {
-    ...base,
-    ...override,
-    image:
-      base.image || override.image
-        ? {
-            ...base.image,
-            ...override.image,
-          }
-        : undefined,
-  };
 }
 
 function matchesProviderScopedModelId(params: {
@@ -1475,76 +1459,6 @@ function shouldCompareProviderRuntimeResolvedModel(params: {
       },
     }) ?? false
   );
-}
-
-function resolveConfiguredFallbackReasoning(params: {
-  provider: string;
-  compat?: unknown;
-  reasoning?: boolean;
-}): boolean {
-  return resolveConfiguredModelReasoning(params) ?? false;
-}
-
-function resolveConfiguredModelReasoning(params: {
-  provider: string;
-  compat?: unknown;
-  reasoning?: boolean;
-}): boolean | undefined {
-  if (params.reasoning !== undefined) {
-    return params.reasoning;
-  }
-  return isVllmQwenThinkingCompat(params) ? true : undefined;
-}
-
-function resolveMergedConfiguredModelReasoning(params: {
-  provider: string;
-  configuredCompat?: unknown;
-  resolvedCompat?: unknown;
-  configuredReasoning?: boolean;
-  discoveredReasoning?: boolean;
-}): boolean {
-  if (params.configuredReasoning !== undefined) {
-    return params.configuredReasoning;
-  }
-  if (isVllmQwenThinkingCompat({ provider: params.provider, compat: params.configuredCompat })) {
-    return true;
-  }
-  return (
-    resolveConfiguredModelReasoning({
-      provider: params.provider,
-      compat: params.resolvedCompat,
-      reasoning: params.discoveredReasoning,
-    }) ?? false
-  );
-}
-
-function isVllmQwenThinkingCompat(params: { provider: string; compat?: unknown }): boolean {
-  const thinkingFormat = readCompatThinkingFormat(params.compat);
-  return (
-    normalizeProviderId(params.provider) === "vllm" &&
-    (thinkingFormat === "qwen" || thinkingFormat === "qwen-chat-template")
-  );
-}
-
-function readCompatThinkingFormat(compat: unknown): string | undefined {
-  if (!compat || typeof compat !== "object" || Array.isArray(compat)) {
-    return undefined;
-  }
-  const thinkingFormat = (compat as { thinkingFormat?: unknown }).thinkingFormat;
-  return typeof thinkingFormat === "string" ? thinkingFormat : undefined;
-}
-
-function mergeModelCompat(
-  base: ModelCompatConfig | undefined,
-  override: ModelCompatConfig | undefined,
-): ModelCompatConfig | undefined {
-  if (!base) {
-    return override;
-  }
-  if (!override) {
-    return base;
-  }
-  return { ...base, ...override };
 }
 
 function normalizeProviderModelRef(params: {
